Solution day 5
Uses a simple recursive function to convert the binary row and column
codes into numbers. It would have been possible to use string replace to
get a binary number string and then parsing it with Integer/parseInt,
but using a simple recursive function lets me practice binary search.

Finding the seat codes is then as simple as mapping over the parsed row
and column numbers and applying the equation for seat codes.

Task 1 simply finds the biggest number in the list. Task 2 sorts the
list of codes, partitions into two at a time with a step value of one
and finds the two seat numbers that are one seat apart.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>