Solution day 6
Read input and split each group into a string.

For task 1 we can simply remove all linebreaks in each group, then find
distinct characters and count them. This makes the number of answers for
each group which is then summed up.

For task 2 we treat each group as a sequence of character sets. Each set
contains a group members answers. Then we simply calculate the
intersections of each group's sets and that is the number of yes answers
for each group. Then we apply + to sum the values.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>