Solution day 3
Uses frequencies function to get frequency of 0s and 1s.

For part 1 the numbers list is pivoted to obtain a list of the first
digits, then the second digits, and so forth. This can then be used to
find the gamma and epsilon rates.

For part 2 we need to use a recursive function to gradually filter down
the numbers list for each of the oxygen and co2 cases. We again use the
frequencies function to get frequency of 0s and 1s, and then apply a
sorting function to satisfy the preference criteria in the task.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>