Day 4 solution
Makes use of set functions (difference) to see if a column or row
matches the drawn numbers. This is reasonably efficient and makes for
more readable code.

Uses a loop recur pattern for each part of the day 4 task. For the first
one it loops until a match is found, and the score can be calculated.

For the second part it keeps track of matching boards and the numbers
drawn to produce the win for those boards. Additionally it removes the
matched boards from the selection of boards to be matched.

Finally when no more numbers can be drawn, or no more boards are
available we simply calculate the score for the last board in the vector
of matched boards.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>