Add day 6 solution
Code contains both a naive super slow solution and a more performant
map/dictionary based solution. The slow solution includes some parallell
mapping to try to speed up the computation, but this was not sufficient.
The inefficiency comes from trying to model each fish as a separate
state entity and thus having to maintain an ever expanding list of
fishes.

A better solution is instead to model the state as a map of the possible
states a fish can be in, and for each state the count of fishes
currently in that state. Then we can simply do some simple moving of
values, and some basic addition for the 6 state.

The map solution uses a transient map to make things more performant
than with persistent maps as we need to mutate the map for each day
iteration. For 256 iterations this will not have much to say, but for
far more iterations this could have an impact.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>