Day 7 solution
For all position ranging from the min and max crab submarine position we
calculate the distance and cost for each crab in the list returning a
flat list of [position crab cost]. We then partition that list by the
position value and then for each sublist sum up the cost using a
reducer. Finally we select the smallest value in the list.

For part 1 we simply use the distance moved as the cost function. For
part 2 progression arithemtic is used to calculate the moving cost for a
given distance.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>