Day 3 solution
This puzzle can be solved using sets and frequencies of letters.
Generally we split the puzzle input into lines of rucksacks.

For the first solution map over the rucksacks, partition each sack into
compartments of two and find the intersect of characters for the two
sack. This way we find a list of items occurring in both compartments
for each rucksack. Then we just map over each item finding its score and
sum up the scores.

For the second solution map over the list of rucksacks and find the
unique items in each sack by turning each sack into a set and then back
to a list. We then turn this list of unique items per sack into groups
of three and find the frequencies of letters for each group. The item
(letter) with a frequency of three is our answer. Then finally map over
each item find its score and sum up.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>