Day 4 solution
Read the input and split into a flat list of numbers. For both part 1
and 2 of the puzzle we are interested in comparing two pairs of number
so we partition the list of numbers into groups of four.

For the first part we then filter down the puzzle input to the
containing pairs by using simple number comparisons. For part two we
also use simple number comparisons.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>