Day 5 solution
This solution required four functions over the input.
The first function calculates the indices of the crates for each line
based on the width of the input text. This helps us with the second
function which parses the input to find the stacks of crates.

Each stack of crates is then a sequence where the first element of the
seq represents the top of the stack.

The third function parses out the moves to do using a regex on each
line.

The final solve function takes the input reduces over the moves using
the stacks as its accumulator and then simply shuffles the crates around
as dictated by moves. For part 1 we simulate the 1 by 1 crate
restriction by reverse the seq of items to move before concatinating it
to the target stack. For the second part we simply move the list as is.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>