Day 6 solution
In this case the input is a single line of characters of signal data so
we only need to slurp up the input file and then start processing
without splitting lines.

For part 1 we want to look at four characters at a time so we partition
the input into lists of four characters, but use a step size of 1 as we
are interested in all sequences of four characters. Then we can use
frequencies on the four letter sequences and check if any of the
frequency maps includes values greather than 1, i.e. if all characters
are distinct.

Part 2 is more or less identical to part 1, but we instead look at 14
consecutive characters.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>