Day 9 solution
Refactor part 1 to use part 2 solution for a generic approach.

Parses input to get all the moves for the head of the rope. For part 1
we construct a rope with 2 knots (head and tail). For part 2 we
construct a rope with 10 knots head and 9 following knots.

Then it is a matter for reducing over the moves and simulating each
successive knot moving or not depending on whether it is touching
(adjacent) to the previous knot. If it is not (pun intended) we have to
move it according to the rules. If it is already adjacent we return it
as is. At the end of each move simulation we record the position of the
tail (the last knot in the rope) in a set. This way we don't record any
positions twice and get the correct count in the end.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>