Day 10 solution
For the first part I use a reducer to run through the instruction and
add a history of the cycle and x register values. For noop instructions
it is enough to add one entry to the history with an incremented cycle
value. For addx instructions we add two history entries, the first
incrementing only the cycle, the second incrementing the cycle and
adding the value to the x register. Finally we multiply each value in
the history, take the values from the 19th in increments  of 40 and add
them together.

For part 2 we repeat the instructions. Then we map all the instructions
and decide whether the cycle draws the pixel or not. We find the column
being drawn by dividing the cycle on the number of pixels and use the
remainder (modulo operator) and then check if this number matches one of
the sprite positions defined by the x register value. Finally we just
draw # or . based on the boolean output of the draw function.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>