Initial work on day 11 solution part 1
Parse each monkey string to get the list of items, an operation
function, and a test function. The simulation seems to be divided into
rounds, each round consisting of turns. For each turn a monkey gets to
go through its list of items and apply the operation, test and giving
the item to another monkey.

The simulation can probably thus be achieved representing each round as
a function that returns the new "state". Using the iterate and take
functions we can then simulate N number of rounds. For each round we
will need to simulate each monkey runnings its action. This can be
achieved using a nested reducer, the outer runnign each monkey's turn,
and the inner running through each item in the monkey's possession.

Signed-off-by: Snorre Magnus Davøen <post@snorre.io>